+++ packages/video_player_avplay/CHANGELOG.md
@@ -1,3 +1,9 @@
+## NEXT
+
+* Check httper_headers pointer when creating media player.
+* Fix state check in SetVolume.
+* Add a case that plays the video of assets to the example.
+
 ## 0.3.2
 
 * [VVC] Add VVC decoder, disable parse for mp4/vvc, create a new vvc decoder to try decode